First commit: Very simple client and server. Only audio packets are being sent.
The client is two threaded: A listening thread (speakers) and a
recording thread (microphone).

The Server is structured as follow:
    - A main thread listens to incoming connection
    - When an incoming connection is accepted, this connection is
    handled by a dedicated thread in a ClientConnection object
    - The ClientConnection object listens for audio packet and push them
    in a RingBuffer
    - This very RingBuffer is then read by another independent thread,
    the Broadcaster, which then hand the audio packets out to the other
    ClientConnections.

Main Issue: While this architecture works fine for 2 people, it is
terrible as soon as a third connection is made. This issue might have
something to do with the way we handle the packets. The packets should
be merged into a single new packet instead of being send one by one.